Fluffy_Push_One
My adjective I want to describe is fluffy, in both the physical soft
sense as well as the romantic sense. To explain the later one think
about a really cute happy couple that is contsantly holding hands, or
cuddles together when the sit together, just cute romantic stuff like
that.

And since I am felling a little sick and pressed for time I was only
able to code some of the assets that program would have. Such as the
teddy bear and heart.